+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This log documents significant changes for each release.
 
+## Not yet released
+
+* #60 When not explicitly set, derive database default fallback locale on the basis 
+of installed reference sets rather than the system default locale
+
 ## [1.4.1292] - 2023-09-19
 
 * #59 Improve error reporting if installed reference sets do not support the requested, or default fallback locale